+++ react-components/src/components/Icons/docs/coloredIcon.tsx
