+++ react-components/src/components/Icons/docs/icon.tsx
