+++ react-components/src/components/Icons/docs/style.module.scss
