+++ react-components/src/components/Icons/icons-list.ts
