+++ react-components/src/components/Icons/stories/ColoredIcons.stories.tsx
