+++ react-components/src/components/Icons/style.module.scss
