+++ react-components/src/components/Spinner/docs/style.module.scss
