+++ react-components/src/components/Spinner/index.tsx
