+++ react-components/src/components/Spinner/style.module.scss
