+++ react-components/src/components/Typography/text.tsx
